Fix #1015: `convertTime` expects to return `int`
`Message::convertTime` expects to return `int`, but PHP `floor()` returns `float` or `boolean` (FALSE).

+++ _protected/framework/Layout/Form/Message.trait.php
@@ -117,7 +117,9 @@ trait Message
     {
         if ($iWaitTime > 60) {
             $iDivide = ($iWaitTime < 1440) ? 60 : 1440;
-            $iWaitTime = floor($iWaitTime / $iDivide);
+
+            // Cast float value from `floor()` to integer
+            $iWaitTime = (int)floor($iWaitTime / $iDivide);
         }
 
         return $iWaitTime;
